refactor: move expandable section logic to partner search component for better organization
